Don't include graphviz.hpp from graphml.hpp (#326)
