+++ include/boost/graph/exception.hpp
@@ -53,6 +53,53 @@ struct BOOST_SYMBOL_VISIBLE not_complete : public bad_graph
     not_complete() : bad_graph("The graph must be complete.") {}
 };
 
+struct BOOST_SYMBOL_VISIBLE graph_exception : public std::exception
+{
+    ~graph_exception() throw() BOOST_OVERRIDE {}
+    const char* what() const throw() BOOST_OVERRIDE = 0;
+};
+
+struct BOOST_SYMBOL_VISIBLE bad_parallel_edge : public graph_exception
+{
+    std::string from;
+    std::string to;
+    mutable std::string statement;
+    bad_parallel_edge(const std::string& i, const std::string& j)
+    : from(i), to(j)
+    {
+    }
+
+    ~bad_parallel_edge() throw() BOOST_OVERRIDE {}
+    const char* what() const throw() BOOST_OVERRIDE
+    {
+        if (statement.empty())
+            statement = std::string("Failed to add parallel edge: (") + from
+                + "," + to + ")\n";
+
+        return statement.c_str();
+    }
+};
+
+struct BOOST_SYMBOL_VISIBLE directed_graph_error : public graph_exception
+{
+    ~directed_graph_error() throw() BOOST_OVERRIDE {}
+    const char* what() const throw() BOOST_OVERRIDE
+    {
+        return "read_graphviz: "
+               "Tried to read a directed graph into an undirected graph.";
+    }
+};
+
+struct BOOST_SYMBOL_VISIBLE undirected_graph_error : public graph_exception
+{
+    ~undirected_graph_error() throw() BOOST_OVERRIDE {}
+    const char* what() const throw() BOOST_OVERRIDE
+    {
+        return "read_graphviz: "
+               "Tried to read an undirected graph into a directed graph.";
+    }
+};
+
 } // namespace boost
 
 #endif // BOOST_GRAPH_EXCEPTION_HPP
